Adds Command Framework (#10)
* Revert "Resolve syntax error in #2 (#7)"

This reverts commit e5a1759cc05fba51097a58f815cfa6a4d0b19d54.

revert: Regression in logging

Undoes the commit that modified line 8 that handles the logging entry point for the bot.

* fix: Change subclass to correct class reference

Changes the MyClient object's subclass from the invalid `discord.bot` to `commands.bot`.

* fix: Reapply commit e5a1759 with the correct logging entry point

Change all references to client to bot in accordance with commit e5a1759 with the appropriate logging entry point.

* refactor: Change logging capitalization

Change `discord.Bot` to `discord.bot` to match with other existing logging entry points.

* feat: Add framework for front end work

Creates a basic extension that acts a skeleton for the front end of the bot. Contributes to resolution of issue #4.